+++ src/main/java/org/cojen/tupl/_LocalDatabase.java
@@ -235,6 +235,9 @@ final class _LocalDatabase extends AbstractDatabase {
 
     private volatile ExecutorService mSorterExecutor;
 
+    // Maps registered cursor ids to index ids.
+    private _Tree mCursorRegistry;
+
     private volatile int mClosed;
     private volatile Throwable mClosedCause;
 
@@ -677,11 +680,13 @@ final class _LocalDatabase extends AbstractDatabase {
                 }
             }
 
+            _Tree cursorRegistry = null;
             if (openMode != OPEN_TEMP) {
                 _Tree tree = openInternalTree(_Tree.FRAGMENTED_TRASH_ID, false, config);
                 if (tree != null) {
                     mFragmentedTrash = new _FragmentedTrash(tree);
                 }
+                cursorRegistry = openInternalTree(_Tree.CURSOR_REGISTRY_ID, false, config);
             }
 
             // Limit maximum non-fragmented entry size to 0.75 of usable node size.
@@ -728,6 +733,21 @@ final class _LocalDatabase extends AbstractDatabase {
                     }
                 }
 
+                LHashTable.Obj<_TreeCursor> cursors = new LHashTable.Obj<>(4);
+                if (cursorRegistry != null) {
+                    Cursor c = cursorRegistry.newCursor(Transaction.BOGUS);
+                    for (c.first(); c.key() != null; c.next()) {
+                        long cursorId = decodeLongBE(c.key(), 0);
+                        long indexId = decodeLongBE(c.value(), 0);
+                        _Tree tree = (_Tree) anyIndexById(indexId);
+                        _TreeCursor cursor = new _TreeCursor(tree);
+                        cursor.mKeyOnly = true;
+                        cursor.mCursorId = cursorId;
+                        cursors.insert(cursorId).value = cursor;
+                    }
+                    cursorRegistry.forceClose();
+                }
+
                 if (mCustomTxnHandler != null) {
                     // Although handler shouldn't access the database yet, be safe and call
                     // this method at the point that the database is mostly functional. All
@@ -756,7 +776,7 @@ final class _LocalDatabase extends AbstractDatabase {
                         }
                     } else {
                         _ReplRedoEngine engine = new _ReplRedoEngine
-                            (rm, config.mMaxReplicaThreads, this, txns);
+                            (rm, config.mMaxReplicaThreads, this, txns, cursors);
                         mRedoWriter = engine.initWriter(redoNum);
 
                         // Cannot start recovery until constructor is finished and final field
@@ -793,7 +813,7 @@ final class _LocalDatabase extends AbstractDatabase {
                             _RedoLog.deleteOldFile(config.mBaseFile, logId - i);
                         }
 
-                        _RedoLogApplier applier = new _RedoLogApplier(this, txns);
+                        _RedoLogApplier applier = new _RedoLogApplier(this, txns, cursors);
                         _RedoLog replayLog = new _RedoLog(config, logId, redoPos);
 
                         // As a side-effect, log id is set one higher than last file scanned.
@@ -2076,6 +2096,12 @@ final class _LocalDatabase extends AbstractDatabase {
                 cursorCount += trash.mTrash.mRoot.countCursors();
             }
 
+            _Tree cursorRegistry = mCursorRegistry;
+            if (cursorRegistry != null) {
+                // Count the cursors which are actively registering cursors. Sounds confusing.
+                cursorCount += cursorRegistry.mRoot.countCursors();
+            }
+
             stats.openIndexes = openTreesCount;
             stats.cursorCount = cursorCount;
 
@@ -2363,6 +2389,13 @@ final class _LocalDatabase extends AbstractDatabase {
             }
         }
 
+        _Tree cursorRegistry = mCursorRegistry;
+        if (cursorRegistry != null) {
+            if (!visitor.apply(cursorRegistry)) {
+                return false;
+            }
+        }
+
         Cursor all = indexRegistryByName().newCursor(null);
         try {
             for (all.first(); all.key() != null; all.next()) {
@@ -2501,6 +2534,10 @@ final class _LocalDatabase extends AbstractDatabase {
                 if (trash != null) {
                     mFragmentedTrash = null;
                     trash.mTrash.forceClose();
+                }
+
+                if (mCursorRegistry != null) {
+                    mCursorRegistry.forceClose();
                 }
 
                 if (mRegistryKeyMap != null) {
@@ -2735,6 +2772,55 @@ final class _LocalDatabase extends AbstractDatabase {
             mRegistry.delete(Transaction.BOGUS, tree.mIdBytes);
         } catch (Throwable e) {
             throw closeOnFailure(this, e);
+        }
+    }
+
+    /**
+     * Should be called before attempting to register a cursor, in case an exception is thrown.
+     */
+    _Tree openCursorRegistry() throws IOException {
+        _Tree cursorRegistry = mCursorRegistry;
+        if (cursorRegistry == null) {
+            mOpenTreesLatch.acquireExclusive();
+            try {
+                if ((cursorRegistry = mCursorRegistry) == null) {
+                    mCursorRegistry = cursorRegistry =
+                        openInternalTree(_Tree.CURSOR_REGISTRY_ID, true);
+                }
+            } finally {
+                mOpenTreesLatch.releaseExclusive();
+            }
+        }
+
+        return cursorRegistry;
+    }
+
+    /**
+     * Should be called after the cursor id has been assigned, with the commit lock held.
+     */
+    void registerCursor(_Tree cursorRegistry, _TreeCursor cursor) throws IOException {
+        try {
+            byte[] cursorIdBytes = new byte[8];
+            encodeLongBE(cursorIdBytes, 0, cursor.mCursorId);
+            cursorRegistry.store(Transaction.BOGUS, cursorIdBytes, cursor.mTree.mIdBytes);
+        } catch (Throwable e) {
+            try {
+                cursor.unregister();
+            } catch (Throwable e2) {
+                suppress(e, e2);
+            }
+            throw e;
+        }
+    }
+
+    void unregisterCursor(_TreeCursor cursor) {
+        try {
+            byte[] cursorIdBytes = new byte[8];
+            encodeLongBE(cursorIdBytes, 0, cursor.mCursorId);
+            openCursorRegistry().store(Transaction.BOGUS, cursorIdBytes, null);
+            cursor.mCursorId = 0;
+        } catch (Throwable e) {
+            // Database is borked, cleanup later.
         }
     }